% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrLASSO.S5.R
\name{rrLASSO.S5}
\alias{rrLASSO.S5}
\title{Bayesian Reciprocal Regularization}
\usage{
rrLASSO.S5(
  X,
  y,
  S = 30,
  lam = 1,
  intercept = TRUE,
  verbose = FALSE,
  seed = 1234
)
}
\arguments{
\item{X}{A numeric matrix with standardized predictors in columns and samples in rows.}

\item{y}{A mean-centered continuous response variable with matching rows with X.}

\item{S}{A screening size of variables. Default is 30.}

\item{lam}{A tuning parameter for the rrLASSO objective function. Default is 1.}

\item{intercept}{If TRUE, intercept is included in the final OLS fit. The default is TRUE.}

\item{verbose}{If TRUE, the function prints the current status of the S5 in each temperature. The default is FALSE.}

\item{seed}{Seed value for reproducibility. Default is 1234.}
}
\value{
A list containing the following components is returned:
\item{hppm}{Index of the highest posterior probablity model.}
\item{marg.prob}{Marginal posterior probablities of the coefficients.}
\item{beta}{OLS coefficients from the final rLASSO model.}
\item{time}{Computation time in seconds.}
}
\description{
Simplified Shotgun Stochastic Search with Screening (S5) algorithm to
solve the reduced reciprocal LASSO, based on a source code  
adapted from `BayesS5` by Minsuk Shin.
}
\examples{
\dontrun{

#########################
# Load Prostate dataset #
#########################

library(ElemStatLearn)
prost<-prostate

###########################################
# Scale data and prepare train/test split #
###########################################

prost.std <- data.frame(cbind(scale(prost[,1:8]),prost$lpsa))
names(prost.std)[9] <- 'lpsa'
data.train <- prost.std[prost$train,]
data.test <- prost.std[!prost$train,]

##################################
# Extract standardized variables #
##################################

y.train   = data.train$lpsa - mean(data.train$lpsa)
y.test <- data.test$lpsa - mean(data.test$lpsa)
x.train = scale(as.matrix(data.train[,1:8], ncol=8))
x.test = scale(as.matrix(data.test[,1:8], ncol=8))

#################################
# Reduced Reciprocal LASSO (S5) #
#################################

rrLasso<- rrLASSO.S5(x.train, y.train)
y.pred.rrLasso<-x.test\%*\%rrLasso$beta[-1]
mean((y.pred.rrLasso - y.test)^2) # Performance on test data

}

}
\keyword{Bayesian}
\keyword{LASSO,}
\keyword{Reciprocal}
\keyword{S5}
\keyword{regularization,}
