% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGFB.R
\docType{data}
\name{TGFB}
\alias{TGFB}
\title{TGFB gene expression colon cancer dataset from Calon et al. (2012).}
\format{
A data frame of gene expression profiles of 172 genes related to TGF-beta 
in n = 262 human patients and the amount of TGF-beta production measured in each patient (y).
}
\usage{
data(TGFB)
}
\description{
Gene expression profiles of 172 genes related to TGF-beta in n = 262 human colon cancer patients.
TGF-beta is one of the key cytokines implicated in various disease processes including colon cancer.
}
\references{
Rossell D, Telesca D (2017). Nonlocal Priors for High-Dimensional Estimation. 
Journal of the American Statistical Association 112(517): 254-265.

Calon et al. (2012). Dependency of Colorectal Cancer On A TGF-beta-driven Program
in Stromal Cells for Metastasis Initiation. Cancer cell 22(5): 571-584.
}
\keyword{datasets}
